Do not implicitly ship releases
This is a leftover from the hybrid model where we had shipit v2 for
everything except product-details, and shipitscript would talk to shipit
v1 to update the status without touching shipit v2. After switching to
v2 only, we didn't update this block. We didn't notice it because the
production details rebuild step is explicitly triggered by shipitscript,
so we never shipped releases to the public without an explicit
shipitscript "signoff".

Fixes #119