frontend: don't filter releases based on branch when guessing build number
The release name, i.e. (product, version, build_number) tuple, must be
unique, so if we filter on branch we can fail at submit time with a
constraint violation.

This shouldn't matter on prod because the version/branch mapping is
unambiguous but on staging this can happen (different github repo
branches, or hg project repos).

An alternative would be to show an error if the product+version was
already used on a different branch.