Update css-loader
4 major versions and not a single breaking change breaking anything, why
can't every dependency be like this. It looks like it was mostly
defaults changing as well as webpack/node requirement going up. I
checked the resulting css and the size is basically the same between
prod and dev so we don't need to fiddle with more settings and we can
trust defaults