Replace mozilla-frontend-infra's Spinner with a raw CircularProgress
mozilla-frontend-infra/components has been archived since 2021, it's
stuck with material-ui 4 (last version is 7). If we want to update, we
have to get rid of it. This is the first step for this.

`Spinner` was a very lightweight wrapper around `CircularProgress`
already, with a bunch of features we don't use. The main difference
between a raw `CircularProgress` and `Spinner` is that it was always put
into a `text-align: center` div when it was in a loading state. This
was mostly unnecessary except for the few spinners that were used to
indicate that a whole page is loading, I just wrapped those myself as
creating a whole component for that sounds very wasteful...