Don't try to massage the `to` prop in `Link`
I just fixed an issue in balrog that was hidden by said massaging
(https://github.com/mozilla-releng/balrog/issues/3547). This is not even
used at all in here because we're always passing a string to `to` and
not using any `state`. But let's not keep a footgun for no reason