Add a way to force getting an authenticated service with `get_service`
There's a list of services that we return as always authenticated. The
queue service isn't part of that list and shouldn't as we do a lot of
anonymous access with it. However, when implementing cancellation, we
need an authenticated queue service to call some methods there. Rather
than making all queue calls throughout shipit authenticated, I simply
added a second parameter to `get_service` that allows forcing the
function to return a service that has got credentials