Add the model for merge automation
`TaskStatus` here maps to taskcluster statuses, minus the difference
between pending/scheduled and failure/exception as those have the same
high level meaning.

Each merge automation is made up of 4 inputs.

- The product it targets (this might be a bit of a misnomer, because it's
running on repositories, not shipit products. But then firefox's merge
automation is running on multiple repositories so I decided to keep it
simple and keep it named product, have one for firefox). Right now the
goal is to support firefox only but this is future proofing because I
know that firefox-ios would also benefit from merge automation.
- The behavior to run. This is the name of the action to run, e.g
`main-to-beta`.
- The revision to run it from. Technically for firefox this doesn't really
matter because AFAIK it will always merge the tip of the branch but this
is where the action will be taken from and that is important for try.
- And finally whether that is a dry run or not.

Then we have a few properties regarding the status of it. When it gets
created, its task ID is null, status is pending. Then as it starts, we
need to record the task id of the action task we just created and update
the status based on the status of both the action task and the resulting
group.

The rest of the fields is caching, to avoid querying HG again since the
data is static and will never change (and querying hg is **slow**), and
to avoid issues if the configuration ever changes.