Add some basic configuration for merge automation
This is slightly annoying because I would really like to avoid having
configuration on the frontend side but this is the first feature to
require that which means we have to figure out how to do it per
environment. The frontend config duplicates the entire thing in 3
different files, instead here I chose to go with the taskgraph looking
way of using `by-env`. This is not perfect, it's very verbose but it
works.

As the comment says, this duplicates some values (repo URLs, version
file path...) with the frontend configuration but said config shouldn't
exist in the first place so I'm fine with that for now.

Do note that this config is a placeholder for now, used for testing.
It's missing a lot of merge behaviors and is using the wrong values for
beta-to-release so we can local test what happens with always-target-tip