Implement the required API to get merge automation working
The one part in here I'm unhappy with is the completion mechanism for
merge automations because it requires someone querying it to update it
which goes against good practice. I did this to avoid having to change
task groups to add a task similar to `mark-as-shipped` to merge
automation and having to uplift it in every tree.

In practice, that status is only ever used from the shipit UI so it'll
never fall out of sync unless the tasks expire before someone visits the
merge automation page again which is never going to happen.

The other alternative would be to hook the worker that currently handles
product details to poll taskcluster and update the status instead.