Bug 1999890 - Fix sentry alerts wrongly firing for 4xx responses (#1813)
Since the switch to connexion 3 (in
71f06c67d48a02ae8b77a526cbeee31195cba5f3), flask has been running within
an ASGI context with connexion doing the translation between ASGI and
WSGI. When we call `abort(400)`, it raises a
`werkzeug.exceptions.BadRequest` which gets translated to
`starlette.exceptions.HTTPException` which then gets translated by the
ASGI layer into a proper 400.

The issue is that we're using sentry with the flask integration, which
sees the exception bubble outside of flask, unhandled (since it's meant
to be handled by the ASGI layer), and reports it to sentry. To fix this,
we can use the `StarletteIntegration` instead and force disable the
`FlaskIntegration` (sentry tries to be smart here and enables it by
default if it can find flask imported, we have to explicitly tell it
not to)