Prevent the creation of releases when we cannot find a decision task (#1814)
It's a very common issue when creating releases to not wait long enough
for the decision task for the given commit to be completed. It results
in a less than useful error about not finding an indexed task, and also
eats the form data. This adds a new API route to check that a decision
task is indexed and adds a check on the new release page for it.

I didn't bother getting the decision task status because it can only
ever be indexed if it succeeded so there's no point in looking at it.